Refactor and consolidate menu code
